% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_cor}
\alias{circ_cor}
\title{Circular Correlation}
\usage{
circ_cor(x, y, na.rm = FALSE, units = c("degrees", "radians"))
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{y}{A numeric vector of angles of the same length as \code{x}.}

\item{na.rm}{Logical value indicating whether pairs with \code{NA} values
should be stripped.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
A numeric value between -1 and 1.
}
\description{
Computes the Jammalamadaka circular correlation coefficient between two
vectors of angles.
}
\examples{
wind_dir_day1 <- c(10, 45, 350, 20)
wind_dir_day2 <- c(15, 50, 355, 25)
circ_cor(wind_dir_day1, wind_dir_day2)

}
