% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_dist}
\alias{circ_dist}
\title{Absolute Circular Distance}
\usage{
circ_dist(x, y, units = c("degrees", "radians"))
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{y}{A numeric vector of reference angles.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
A numeric vector of absolute distances \verb{[0, 180]} or \verb{[0, pi]}.
}
\description{
Calculates the absolute shortest distance between two angles.
}
\examples{
circ_dist(10, 350) # Returns 20, not -20 or 340

}
