% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_mean}
\alias{circ_mean}
\title{Circular Mean}
\usage{
circ_mean(x, na.rm = FALSE, units = c("degrees", "radians"))
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{na.rm}{Logical value indicating whether \code{NA} values should be stripped.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
The mean angle in the specified units.
}
\description{
Computes the average angle of a vector of circular data.
}
\examples{
circ_mean(c(350, 10))
circ_mean(c(0, 90, 180, NA), na.rm = TRUE)

}
