% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_resultant_length}
\alias{circ_resultant_length}
\title{Mean Resultant Length}
\usage{
circ_resultant_length(x, na.rm = FALSE, units = c("degrees", "radians"))
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{na.rm}{Logical value indicating whether \code{NA} values should be stripped.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
A numeric value between 0 (uniformly distributed) and 1 (perfectly concentrated).
}
\description{
Computes the mean resultant length (often denoted as \emph{R}) of circular data.
This is a measure of concentration for circular distributions.
}
\examples{
circ_resultant_length(c(90, 90, 90)) # 1
circ_resultant_length(c(0, 90, 180, 270)) # 0

}
