% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_var}
\alias{circ_var}
\title{Circular Variance}
\usage{
circ_var(x, na.rm = FALSE, units = c("degrees", "radians"))
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{na.rm}{Logical value indicating whether \code{NA} values should be stripped.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
A numeric value between 0 (no variance) and 1 (maximum variance).
}
\description{
Computes the Mardia circular variance, calculated as 1 minus the mean
resultant length.
}
\examples{
circ_var(c(5, 10, 15))
circ_var(c(0, 90, 180, 270))

}
