% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{circ_wrap}
\alias{circ_wrap}
\title{Wrap Angles to standard circular boundaries}
\usage{
circ_wrap(
  x,
  bounds = c("positive", "centered"),
  units = c("degrees", "radians")
)
}
\arguments{
\item{x}{A numeric vector of angles.}

\item{bounds}{A character string specifying the target range.
\code{"positive"} constrains to \verb{[0, 360)} or \verb{[0, 2pi)}.
\code{"centered"} constrains to \verb{[-180, 180)} or \verb{[-pi, pi)}.}

\item{units}{A character string specifying the unit of measurement.
Must be one of \code{"degrees"} (default) or \code{"radians"}.}
}
\value{
A numeric vector of normalized angles.
}
\description{
Constrains messy or out-of-bounds angles into a standard single rotation.
}
\examples{
circ_wrap(c(370, -45), bounds = "positive")
circ_wrap(c(370, 315), bounds = "centered")

}
