% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{degrad}
\alias{degrad}
\alias{deg2rad}
\alias{rad2deg}
\title{Conversion between degrees and radians}
\usage{
deg2rad(x)

rad2deg(x)
}
\arguments{
\item{x}{A numeric vector of angles.}
}
\value{
A numeric vector of converted values.
}
\description{
Helper functions to convert between degrees and radians.
}
\examples{
deg2rad(47)
deg2rad(c(10, 30, 40))
rad2deg(pi / 7)

}
